Configured permissions for Project Invite code management via web token

+++ llmur/src/routes/project_invite_code.rs
@@ -1,10 +1,12 @@
 use crate::data::project::{ProjectId, ProjectRole};
 use crate::data::project_invite_code::{ProjectInviteCode, ProjectInviteCodeId};
 use crate::data::utils::current_timestamp_s;
-use crate::errors::{AuthorizationError, LLMurError};
-use crate::routes::middleware::user_context::{AuthorizationManager, UserContext, UserContextExtractionResult};
+use crate::errors::{AuthorizationError, DataAccessError, LLMurError};
 use crate::routes::StatusResponse;
-use crate::{impl_from_vec_result, LLMurState};
+use crate::routes::middleware::user_context::{
+    AuthorizationManager, UserContext, UserContextExtractionResult,
+};
+use crate::{LLMurState, impl_from_vec_result};
 use axum::extract::{Path, State};
 use axum::routing::{delete, post};
 use axum::{Extension, Json, Router};
@@ -19,10 +21,7 @@ pub(crate) fn routes(state: Arc<LLMurState>) -> Router<Arc<LLMurState>> {
         .with_state(state.clone())
 }
 
-#[tracing::instrument(
-    name = "handler.create.project_invite_code",
-    skip(state, ctx, payload)
-)]
+#[tracing::instrument(name = "handler.create.project_invite_code", skip(state, ctx, payload))]
 pub(crate) async fn create_project_invite_code(
     Extension(ctx): Extension<UserContextExtractionResult>,
     State(state): State<Arc<LLMurState>>,
@@ -30,45 +29,25 @@ pub(crate) async fn create_project_invite_code(
 ) -> Result<Json<GetProjectInviteCodeResult>, LLMurError> {
     let user_context = ctx.require_authenticated_user()?;
 
-    match user_context {
-        UserContext::MasterUser => {
-            let invite = state.data.create_invite_code(
-                &payload.project_id,
-                &payload.role.unwrap_or(ProjectRole::Guest),
-                &payload.validity,
-                &payload.code_length,
-                &state.metrics
-            ).await?;
-
-            Ok(Json(invite.into()))
-        }
-        UserContext::WebAppUser { user, .. } => {
-            /*
-            let memberships = state.data.projects()
-                .get_project_memberships(&payload.project_id)
-                .await?
-                .ok_or(LLMurError::AdminResourceNotFound)?;
-
-            if let Some(membership) = memberships.iter().find(|m| m.user_id == user.id) {
-                if membership.role == ProjectRole::Admin {
-                    let result = state.data.invites().create_invite(
-                        &payload.project_id,
-                        &payload.role.unwrap_or(ProjectRole::Guest),
-                        &payload.validity,
-                        &payload.code_length,
-                    ).await?;
-                    return Ok(Json(StatusResponse {
-                        success: true,
-                        message: Some(format!("Invite code {} created successfully", result.id)),
-                    }));
-                }
-            }
-
-             */
-
-            Err(AuthorizationError::AccessDenied)?
-        }
+    if !user_context
+        .has_project_admin_access(state.clone(), &payload.project_id)
+        .await?
+    {
+        return Err(AuthorizationError::AccessDenied)?;
     }
+
+    let invite = state
+        .data
+        .create_invite_code(
+            &payload.project_id,
+            &payload.role.unwrap_or(ProjectRole::Guest),
+            &payload.validity,
+            &payload.code_length,
+            &state.metrics,
+        )
+        .await?;
+
+    Ok(Json(invite.into()))
 }
 
 #[tracing::instrument(
@@ -85,42 +64,22 @@ pub(crate) async fn delete_project_invite_code(
 ) -> Result<Json<StatusResponse>, LLMurError> {
     let user_context = ctx.require_authenticated_user()?;
 
+    let project_invite_code = state
+        .data
+        .get_invite_code(&id, &state.metrics)
+        .await?
+        .ok_or(DataAccessError::ResourceNotFound)?;
 
-    match user_context {
-        UserContext::MasterUser => {
-            let result = state.data.delete_invite_code(&id, &state.metrics).await?;
-
-            Ok(Json(StatusResponse {
-                success: result != 0,
-                message: None,
-            }))
-        }
-        UserContext::WebAppUser { user, .. } => {
-            /*
-            let maybe_invite = state.data.invites().get_invite_with_id(&id).await?;
-
-            if let Some(invite) = maybe_invite {
-                let memberships = state.data.projects()
-                    .get_project_memberships(&invite.project_id)
-                    .await?
-                    .ok_or(LLMurError::AdminResourceNotFound)?;
-
-                if let Some(membership) = memberships.iter().find(|m| m.user_id == user.id) {
-                    if membership.role == ProjectRole::Admin {
-                        let result = state.data.invites().delete_invite(&id).await?;
-                        return Ok(Json(StatusResponse {
-                            success: result != 0,
-                            message: None,
-                        }));
-                    }
-                }
-            }
-
-             */
-
-            Err(AuthorizationError::AccessDenied)?
-        }
+    if !user_context.has_project_admin_access(state.clone(), &project_invite_code.project_id).await? {
+        return Err(AuthorizationError::AccessDenied)?;
     }
+
+    let result = state.data.delete_invite_code(&id, &state.metrics).await?;
+
+    Ok(Json(StatusResponse {
+        success: result != 0,
+        message: None,
+    }))
 }
 // endregion: --- Routes
 
@@ -146,10 +105,14 @@ pub(crate) struct GetProjectInviteCodeResult {
 #[derive(Serialize)]
 pub(crate) struct ListProjectInviteCodesResult {
     pub(crate) codes: Vec<GetProjectInviteCodeResult>,
-    pub(crate) total: usize
+    pub(crate) total: usize,
 }
 
-impl_from_vec_result!(GetProjectInviteCodeResult, ListProjectInviteCodesResult, codes);
+impl_from_vec_result!(
+    GetProjectInviteCodeResult,
+    ListProjectInviteCodesResult,
+    codes
+);
 
 impl From<ProjectInviteCode> for GetProjectInviteCodeResult {
     fn from(value: ProjectInviteCode) -> Self {
@@ -157,7 +120,10 @@ impl From<ProjectInviteCode> for GetProjectInviteCodeResult {
             id: value.id,
             project_id: value.project_id,
             code: value.code,
-            valid: value.valid_until.map(|vu| vu  < current_timestamp_s()).unwrap_or(true),
+            valid: value
+                .valid_until
+                .map(|vu| vu < current_timestamp_s())
+                .unwrap_or(true),
             valid_until: value.valid_until,
             role: value.assign_role,
         }
